+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased](https://github.com/rajatthareja/ReportBuilder/compare/v1.3...master)
+### Fixed
+- Config options Fixed
+- Document Fixed
+
+### Added
+- Method to set single option 
 
 ## [1.3](https://github.com/rajatthareja/ReportBuilder/compare/v1.2...v1.3) - 2017-10-22
 ### Changed